third_party/blink/web_tests/external/wpt/geolocation-API/getCurrentPosition_permission_deny.https.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>Geolocation Test: getCurrentPosition location access denied</title>
<link rel="help" href="http://www.w3.org/TR/geolocation-API/#privacy_for_uas" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
  promise_test(async (t) => {
    t.add_cleanup(() => {
      return test_driver.set_permission({ name: "geolocation" }, "prompt");
    });
    await test_driver.set_permission({ name: "geolocation" }, "denied");
    const result = await new Promise((resolve, reject) => {
      navigator.geolocation.getCurrentPosition(reject, resolve);
    });
    assert_true(
      result instanceof GeolocationPositionError,
      "should be a GeolocationPositionError"
    );
    assert_equals(result.code, result.PERMISSION_DENIED);
  }, "User denies access, check that error callback is called with correct code");
</script>
